Refactor imports for @hedvig-ui/InfoRow
