+++ shared/hedvig-ui/InfoRow/info-row.tsx
