+++ src/components/claims/claim-details/components/ClaimPayments.tsx
@@ -11,10 +11,9 @@ import { useClaimPaymentsQuery } from 'api/generated/graphql'
 import { format, parseISO } from 'date-fns'
 import { Spinner } from 'hedvig-ui/sipnner'
 
-import { CardContent } from '@hedvig-ui'
+import { CardContent, InfoRow, InfoTag, InfoText } from '@hedvig-ui'
 import { PaperTitle } from 'components/claims/claim-details/components/claim-items/PaperTitle'
 import { StandaloneMessage } from 'hedvig-ui/animations/standalone-message'
-import { InfoRow, InfoTag, InfoText } from 'hedvig-ui/info-row'
 import { Paragraph, Shadowed, ThirdLevelHeadline } from 'hedvig-ui/typography'
 import React from 'react'
 import { BugFill } from 'react-bootstrap-icons'
