+++ src/components/member/tabs/contracts-tab/agreement/AgreementInfo.tsx
@@ -1,8 +1,8 @@
 import styled from '@emotion/styled'
+import { InfoContainer, InfoRow, InfoText } from '@hedvig-ui'
 import { IconButton } from '@material-ui/core'
 import { Contract, GenericAgreement } from 'api/generated/graphql'
 import { EditStreetInput } from 'components/member/tabs/contracts-tab/agreement/EditStreetInput'
-import { InfoContainer, InfoRow, InfoText } from 'hedvig-ui/info-row'
 import { Paragraph } from 'hedvig-ui/typography'
 import React, { useState } from 'react'
 import { PencilFill } from 'react-bootstrap-icons'
