+++ src/components/member/tabs/contracts-tab/contract/index.tsx
@@ -1,11 +1,16 @@
 import styled from '@emotion/styled'
-import { Card, CardsWrapper } from '@hedvig-ui'
+import {
+  Card,
+  CardsWrapper,
+  InfoContainer,
+  InfoRow,
+  InfoText,
+} from '@hedvig-ui'
 import { Contract as ContractType } from 'api/generated/graphql'
 import { Agreement } from 'components/member/tabs/contracts-tab/agreement'
 import { AgreementsTable } from 'components/member/tabs/contracts-tab/agreement/AgreementsTable'
 import { MasterInception } from 'components/member/tabs/contracts-tab/contract/master-inception'
 import { TerminationDate } from 'components/member/tabs/contracts-tab/contract/termination-date'
-import { InfoContainer, InfoRow, InfoText } from 'hedvig-ui/info-row'
 import { ThirdLevelHeadline } from 'hedvig-ui/typography'
 import React from 'react'
 import { getSignSource } from 'utils/contract'
