+++ src/components/member/tabs/contracts-tab/trial/index.tsx
@@ -1,7 +1,13 @@
 import styled from '@emotion/styled'
-import { ButtonLink, Card, CardsWrapper } from '@hedvig-ui'
+import {
+  ButtonLink,
+  Card,
+  CardsWrapper,
+  InfoContainer,
+  InfoRow,
+  InfoText,
+} from '@hedvig-ui'
 import { Trial } from 'api/generated/graphql'
-import { InfoContainer, InfoRow, InfoText } from 'hedvig-ui/info-row'
 import { Paragraph, ThirdLevelHeadline } from 'hedvig-ui/typography'
 import React from 'react'
 import { convertEnumToTitle } from 'utils/text'
