+++ src/components/member/tabs/debt-tab/index.tsx
@@ -1,4 +1,10 @@
-import { Card, CardsWrapper } from '@hedvig-ui'
+import {
+  Card,
+  CardsWrapper,
+  InfoContainer,
+  InfoRow,
+  InfoText,
+} from '@hedvig-ui'
 import { WhitelistMemberButton } from 'components/member/tabs/debt-tab/WhitelistMemberButton'
 import { RefreshButton } from 'components/member/tabs/shared/refresh-button'
 import { useGetPerson } from 'graphql/use-get-person'
@@ -7,7 +13,6 @@ import {
   LoadingMessage,
   StandaloneMessage,
 } from 'hedvig-ui/animations/standalone-message'
-import { InfoContainer, InfoRow, InfoText } from 'hedvig-ui/info-row'
 import { FlagOrbIndicator } from 'hedvig-ui/orb-indicator'
 import { Spacing } from 'hedvig-ui/spacing'
 import { MainHeadline } from 'hedvig-ui/typography'
