+++ src/features/tools/campaign-codes/components/CreateCampaignCode.tsx
@@ -1,7 +1,6 @@
-import { getTextFromEnumValue } from '@hedvig-ui'
+import { getTextFromEnumValue, InfoContainer } from '@hedvig-ui'
 import { VisibleNoDiscountForm } from 'features/tools/campaign-codes/forms/VisibleNoDiscountForm'
 import { CreatableIncentiveTypes } from 'features/tools/campaign-codes/utils'
-import { InfoContainer } from 'hedvig-ui/info-row'
 import { SearchableDropdown } from 'hedvig-ui/searchable-dropdown'
 import { Spacing } from 'hedvig-ui/spacing'
 import { ThirdLevelHeadline } from 'hedvig-ui/typography'
